Add files via upload
A project (still in progress) aimed at identifying which sex and age groups in the United States are the riskiest from an auto insurer's perspective in determining how much and how often they tend to request claims. This code is heavily focused on finding an appropriate model for predicting the claim amount of an individual policyholder.